MOBI-12 reworded large and small screen sized detection for tablets/handhelds. Also integraed new ipad and honeycomb themes.
git-svn-id: http://anonsvn.icesoft.org/repo/icemobile/trunk/icemobile@26499 1de00972-ac04-4911-a46b-06793c8d78d9

+++ samples/mobileshowcase/src/main/java/org/icemobile/samples/mobileshowcase/view/examples/input/input/InputBean.java
@@ -65,7 +65,7 @@ public class InputBean extends ExampleImpl<InputBean> implements
     private String date;
     private String textArea;
     private String urlInput;
-    private int numberInput;
+    private double numberInput;
     private Date dateInput;
 
 
@@ -89,11 +89,11 @@ public class InputBean extends ExampleImpl<InputBean> implements
         this.passwordInput = passWordInput;
     }
 
-    public int getNumberInput() {
+    public double getNumberInput() {
         return numberInput;
     }
 
-    public void setNumberInput(int numberInput) {
+    public void setNumberInput(double numberInput) {
         this.numberInput = numberInput;
     }
